+++ tests/test_data/sample-3-classes-real-location.sh
@@ -1,2 +1,4 @@
-autoclass -search sample-3-classes-real-location.db2 sample-3-classes-real-location.hd2 sample-3-classes-real-location.model sample-3-classes-real-location.s-params 
+autoclass -search sample-3-classes-real-location.db2 sample-3-classes-real-location.hd2 sample-3-classes-real-location.model sample-3-classes-real-location.s-params <<EOF
+y
+EOF
 autoclass -reports sample-3-classes-real-location.results-bin sample-3-classes-real-location.search sample-3-classes-real-location.r-params