+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add dendrogram of classes
 - add reproducible run option (test only)
 - add tests for Input() class
+- add tests for Output() class
 
 **0.1.12**
 - update write_cluster_stats() methods with mean/std column